+++ src/promptflow/promptflow/_sdk/_load_functions.py
@@ -68,7 +68,6 @@ def load_common(
 
 def load_flow(
     source: Union[str, PathLike, IO[AnyStr]],
-    is_async_call: Optional[bool] = None,
     **kwargs,
 ) -> Flow:
     """Load flow from YAML file.
@@ -77,13 +76,10 @@ def load_flow(
         If the source is a path, it will be open and read.
         An exception is raised if the file does not exist.
     :type source: Union[PathLike, str]
-    :param is_async_call: Optional argument to indicate the return value is an async function.
-        If True, the return value is an async function, otherwise, it is a sync function.
-    :type is_async_call: bool
     :return: A Flow object
     :rtype: Flow
     """
-    return Flow.load(source, is_async_call=is_async_call, **kwargs)
+    return Flow.load(source, **kwargs)
 
 
 def load_run(
