+++ src/promptflow/promptflow/_sdk/_submitter/experiment_orchestrator.py
@@ -54,7 +54,7 @@ from promptflow._sdk._submitter.utils import (
 from promptflow._sdk._utils import overwrite_null_std_logger
 from promptflow._sdk.entities import Run
 from promptflow._sdk.entities._experiment import Experiment, ExperimentTemplate
-from promptflow._sdk.entities._flow import ProtectedFlow
+from promptflow._sdk.entities._flow import Flow
 from promptflow._sdk.operations import RunOperations
 from promptflow._sdk.operations._local_storage_operations import LocalStorageOperations
 from promptflow._utils.inputs_mapping_utils import apply_inputs_mapping
@@ -102,7 +102,7 @@ class ExperimentOrchestrator:
             node
             for node in template.nodes
             if node.type == ExperimentNodeType.FLOW
-            and ProtectedFlow._get_flow_definition(node.path) == ProtectedFlow._get_flow_definition(flow_path)
+            and Flow._get_flow_definition(node.path) == Flow._get_flow_definition(flow_path)
         ]
         if not start_nodes:
             raise ExperimentValueError(f"Flow {flow_path.as_posix()} not found in experiment {template.dir_name!r}.")
