+++ src/promptflow/promptflow/_sdk/_submitter/test_submitter.py
@@ -13,7 +13,7 @@ from colorama import Fore, init
 from promptflow._internal import ConnectionManager
 from promptflow._sdk._constants import PROMPT_FLOW_DIR_NAME
 from promptflow._sdk._utils import dump_flow_result, parse_variant
-from promptflow._sdk.entities._flow import FlowBase, FlowContext, ProtectedFlow
+from promptflow._sdk.entities._flow import Flow, FlowBase, FlowContext
 from promptflow._sdk.operations._local_storage_operations import LoggerOperations
 from promptflow._utils.context_utils import _change_working_dir
 from promptflow._utils.exception_utils import ErrorResponse
@@ -30,7 +30,7 @@ from ..._utils.dataclass_serializer import convert_eager_flow_output_to_dict
 from ..._utils.logger_utils import get_cli_sdk_logger
 from ...batch import APIBasedExecutorProxy, CSharpExecutorProxy
 from .._configuration import Configuration
-from ..entities._eager_flow import EagerFlow
+from ..entities._eager_flow import FlexFlow
 from .utils import (
     SubmitterHelper,
     print_chat_output,
@@ -62,12 +62,12 @@ class TestSubmitter:
 
     def __init__(
         self,
-        flow: Union[ProtectedFlow, EagerFlow],
+        flow: Union[Flow, FlexFlow],
         flow_context: FlowContext,
         client=None,
     ):
         self._flow = flow
-        self.entry = flow.entry if isinstance(flow, EagerFlow) else None
+        self.entry = flow.entry if isinstance(flow, FlexFlow) else None
         self._origin_flow = flow
         self._dataplane_flow = None
         self.flow_context = flow_context
@@ -162,7 +162,7 @@ class TestSubmitter:
     def _resolve_connections(cls, flow: FlowBase, client):
         if flow.language == FlowLanguage.CSharp:
             # TODO: check if this is a shared logic
-            if isinstance(flow, EagerFlow):
+            if isinstance(flow, FlexFlow):
                 # connection overrides are not supported for eager flow for now
                 return {}
 
@@ -183,7 +183,7 @@ class TestSubmitter:
         raise UserErrorException(f"Unsupported flow language {flow.language}")
 
     @classmethod
-    def _resolve_environment_variables(cls, environment_variable_overrides, flow: ProtectedFlow, client):
+    def _resolve_environment_variables(cls, environment_variable_overrides, flow: Flow, client):
         return SubmitterHelper.load_and_resolve_environment_variables(
             flow=flow, environment_variable_overrides=environment_variable_overrides, client=client
         )
