+++ src/promptflow/promptflow/_sdk/_utils.py
@@ -5,7 +5,6 @@ import collections
 import datetime
 import hashlib
 import json
-import multiprocessing
 import os
 import platform
 import re
@@ -37,8 +36,6 @@ from promptflow._sdk._constants import (
     AZURE_WORKSPACE_REGEX_FORMAT,
     DAG_FILE_NAME,
     DEFAULT_ENCODING,
-    FLOW_META_JSON,
-    FLOW_META_JSON_GEN_TIMEOUT,
     FLOW_TOOLS_JSON,
     FLOW_TOOLS_JSON_GEN_TIMEOUT,
     HOME_PROMPT_FLOW_DIR,
@@ -59,7 +56,6 @@ from promptflow._sdk._constants import (
 )
 from promptflow._sdk._errors import (
     DecryptConnectionError,
-    GenerateFlowMetaJsonError,
     GenerateFlowToolsJsonError,
     StoreConnectionEncryptionKeyError,
     UnsecureConnectionError,
@@ -71,6 +67,7 @@ from promptflow._utils.logger_utils import get_cli_sdk_logger
 from promptflow._utils.utils import _match_reference
 from promptflow._utils.yaml_utils import dump_yaml, load_yaml, load_yaml_string
 from promptflow.contracts.tool import ToolType
+from promptflow.core._utils import generate_flow_meta as _generate_flow_meta
 from promptflow.exceptions import ErrorTarget, UserErrorException
 
 logger = get_cli_sdk_logger()
@@ -1143,86 +1140,6 @@ def _generate_meta_from_file(working_dir, source_path, entry, meta_dict, excepti
             exception_list.append(str(e))
 
 
-def _generate_flow_meta(
-    flow_directory: Path,
-    source_path: str,
-    entry: str,
-    timeout: int,
-    *,
-    load_in_subprocess: bool = True,
-) -> Dict[str, dict]:
-    """Generate tool meta from files.
-
-    :param flow_directory: flow directory
-    :param tools: tool list
-    :param raise_error: whether raise error when generate meta failed
-    :param timeout: timeout for generate meta
-    :param include_errors_in_output: whether include errors in output
-    :param load_in_subprocess: whether load tool meta with subprocess to prevent system path disturb. Default is True.
-        If set to False, will load tool meta in sync mode and timeout need to be handled outside current process.
-    :return: tool meta dict
-    """
-    if load_in_subprocess:
-        # use multiprocess generate to avoid system path disturb
-        manager = multiprocessing.Manager()
-        meta_dict = manager.dict()
-        exception_list = manager.list()
-        p = multiprocessing.Process(
-            target=_generate_meta_from_file, args=(flow_directory, source_path, entry, meta_dict, exception_list)
-        )
-        p.start()
-        p.join(timeout=timeout)
-        if p.is_alive():
-            logger.warning(f"Generate meta timeout after {timeout} seconds, terminate the process.")
-            p.terminate()
-            p.join()
-    else:
-        meta_dict, exception_list = {}, []
-
-        #  There is no built-in method to forcefully stop a running thread/coroutine in Python
-        #  because abruptly stopping a thread can cause issues like resource leaks,
-        #  deadlocks, or inconsistent states.
-        #  Caller needs to handle the timeout outside current process.
-        logger.warning(
-            "Generate meta in current process and timeout won't take effect. "
-            "Please handle timeout manually outside current process."
-        )
-        _generate_meta_from_file(flow_directory, source_path, entry, meta_dict, exception_list)
-    # directly raise error if failed to generate meta
-    if len(exception_list) > 0:
-        error_message = "Generate meta failed, detail error:\n" + str(exception_list)
-        raise GenerateFlowMetaJsonError(error_message)
-    return dict(meta_dict)
-
-
-def generate_flow_meta(
-    flow_directory: Union[str, Path],
-    source_path: str,
-    entry: str,
-    dump: bool = True,
-    timeout: int = FLOW_META_JSON_GEN_TIMEOUT,
-    load_in_subprocess: bool = True,
-) -> dict:
-    """Generate flow.json for a flow directory."""
-
-    flow_meta = _generate_flow_meta(
-        flow_directory=flow_directory,
-        source_path=source_path,
-        entry=entry,
-        timeout=timeout,
-        load_in_subprocess=load_in_subprocess,
-    )
-
-    if dump:
-        # dump as flow.tools.json
-        promptflow_folder = flow_directory / PROMPT_FLOW_DIR_NAME
-        promptflow_folder.mkdir(exist_ok=True)
-        with open(promptflow_folder / FLOW_META_JSON, mode="w", encoding=DEFAULT_ENCODING) as f:
-            json.dump(flow_meta, f, indent=4)
-
-    return flow_meta
-
-
 def extract_workspace_triad_from_trace_provider(trace_provider: str) -> AzureMLWorkspaceTriad:
     match = re.match(AZURE_WORKSPACE_REGEX_FORMAT, trace_provider)
     if not match or len(match.groups()) != 5:
@@ -1244,3 +1161,6 @@ def overwrite_null_std_logger():
         sys.stdout = open(os.devnull, "w")
     if sys.stderr is None:
         sys.stderr = sys.stdout
+
+
+generate_flow_meta = _generate_flow_meta
