+++ src/promptflow/promptflow/azure/operations/_flow_operations.py
@@ -244,7 +244,7 @@ class FlowOperations(WorkspaceTelemetryMixin, _ScopeDependentOperations):
     @staticmethod
     def _validate_flow_schema(source, display_name=None, type=None, **kwargs):
         """Validate the flow schema."""
-        from promptflow._sdk.entities._flow import ProtectedFlow
+        from promptflow._sdk.entities._flow import Flow
 
         params_override = copy.deepcopy(kwargs)
         if display_name is not None:
@@ -252,7 +252,7 @@ class FlowOperations(WorkspaceTelemetryMixin, _ScopeDependentOperations):
         if type is not None:
             params_override["type"] = type
 
-        flow_entity = ProtectedFlow.load(source=source, params_override=params_override)
+        flow_entity = Flow.load(source=source, params_override=params_override)
         flow_entity._validate(raise_error=True)  # raise error if validation failed
         flow_dict = flow_entity._dump_for_validation()
         return flow_dict
@@ -601,10 +601,10 @@ class FlowOperations(WorkspaceTelemetryMixin, _ScopeDependentOperations):
     @classmethod
     def _generate_meta_for_eager_flow(cls, code):
         from promptflow import load_flow as load_local_flow
-        from promptflow._sdk.entities._eager_flow import EagerFlow
+        from promptflow._sdk.entities._eager_flow import FlexFlow
 
         flow = load_local_flow(code.path)
-        if isinstance(flow, EagerFlow) and flow.language == FlowLanguage.Python:
+        if isinstance(flow, FlexFlow) and flow.language == FlowLanguage.Python:
             # TODO: support generate meta for CSharp flow
             generate_flow_meta(
                 flow_directory=code.path,
