+++ src/promptflow/tests/sdk_cli_test/unittests/test_flow.py
@@ -7,8 +7,8 @@ import pytest
 from marshmallow import ValidationError
 
 from promptflow import load_flow
-from promptflow._sdk.entities._eager_flow import EagerFlow
-from promptflow._sdk.entities._flow import ProtectedFlow
+from promptflow._sdk.entities._eager_flow import FlexFlow
+from promptflow._sdk.entities._flow import Flow
 
 FLOWS_DIR = Path("./tests/test_configs/flows")
 EAGER_FLOWS_DIR = Path("./tests/test_configs/eager_flows")
@@ -26,7 +26,7 @@ class TestRun:
     )
     def test_eager_flow_load(self, kwargs):
         flow = load_flow(**kwargs)
-        assert isinstance(flow, EagerFlow)
+        assert isinstance(flow, FlexFlow)
 
     @pytest.mark.parametrize(
         "kwargs",
@@ -37,11 +37,11 @@ class TestRun:
     )
     def test_dag_flow_load(self, kwargs):
         flow = load_flow(**kwargs)
-        assert isinstance(flow, ProtectedFlow)
+        assert isinstance(flow, Flow)
 
     def test_flow_load_advanced(self):
         flow = load_flow(source=EAGER_FLOWS_DIR / "flow_with_environment")
-        assert isinstance(flow, EagerFlow)
+        assert isinstance(flow, FlexFlow)
         assert flow._data["environment"] == {"python_requirements_txt": "requirements.txt"}
 
     @pytest.mark.parametrize(
